+++ Project1/DeleteSetButton.cpp
@@ -7,9 +7,9 @@
 void DeleteSetButton::Activate()
 {
 	ToolkitUI& thisUI = *(ToolkitUI*)(WControl::AllUI()[UIType::ToolkitPage]);
-	if (WControl::allRoomPrefabs().size() > 1) {
+	if (WControl::allRoomPrefabs().size() > 1&& !WorldControl::allRoomPrefabs()[thisUI.choosingSet].first.second) {
 		WControl::allRoomPrefabs().erase(thisUI.choosingSet);
-		WControl::RefreshRoomPrefrab();
+		Load::RefreshRoomPrefrab();
 		std::filesystem::remove_all("Rooms\\" + thisUI.choosingSet);
 		thisUI.ChangeSet(WControl::allRoomPrefabs().begin()->first);
 		thisUI.updateText();
